+++ server/src/interfaces/auth/LoginUserDTO.ts
