+++ server/src/interfaces/auth/RegisterUserDTO.ts
