+++ server/src/models/user.model.ts
@@ -1,5 +1,5 @@
 import { Schema, model } from 'mongoose';
-import {IUser} from '../../interfaces/IUser';
+import {IUser} from '../interfaces/IUser';
 
 const userSchema = new Schema(
     {
